Move logging code to 'simpleflow.log' module for clarity and better imports
